requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aiohttp==3.8.0; python_version >= '3.6'
aiosignal==1.2.0; python_version >= '3.6'
aniso8601==9.0.1
async-timeout==4.0.0; python_version >= '3.6'
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
azure-core==1.20.1
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7; python_full_version >= '3.5.0'
click==8.0.3; python_version >= '3.6'
cryptography==35.0.0; python_version >= '3.6'
flask-restful==0.3.9
flask==2.0.2
frozenlist==1.2.0; python_version >= '3.6'
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==3.3; python_version >= '3'
isodate==0.6.0
itsdangerous==2.0.1; python_version >= '3.6'
jinja2==3.0.2; python_version >= '3.6'
line-bot-sdk==2.0.1
markupsafe==2.0.1; python_version >= '3.6'
msrest==0.6.21
multidict==5.2.0; python_version >= '3.6'
oauthlib==3.1.1; python_version >= '3.6'
pycparser==2.21; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytz==2021.3
requests-oauthlib==1.3.0
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==3.10.0.2
urllib3==1.26.7; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.2; python_version >= '3.6'
yarl==1.7.2; python_version >= '3.6'
